+++ indexer/services/comlink/src/controllers/api/v4/historical-block-trading-rewards-controller.ts
@@ -17,7 +17,6 @@ import { getReqRateLimiter } from '../../../caches/rate-limiters';
 import config from '../../../config';
 import { handleControllerError } from '../../../lib/helpers';
 import { rateLimiterMiddleware } from '../../../lib/rate-limit';
-import { rejectRestrictedCountries } from '../../../lib/restrict-countries';
 import { CheckHistoricalBlockTradingRewardsSchema } from '../../../lib/validation/schemas';
 import { handleValidationErrors } from '../../../request-helpers/error-handler';
 import ExportResponseCodeStats from '../../../request-helpers/export-response-code-stats';
@@ -55,7 +54,6 @@ class HistoricalBlockTradingRewardsController extends Controller {
 
 router.get(
   '/:address',
-  rejectRestrictedCountries,
   rateLimiterMiddleware(getReqRateLimiter),
   ...CheckHistoricalBlockTradingRewardsSchema,
   handleValidationErrors,
