+++ indexer/services/comlink/src/controllers/api/v4/historical-trading-reward-aggregations-controller.ts
@@ -18,7 +18,6 @@ import { getReqRateLimiter } from '../../../caches/rate-limiters';
 import config from '../../../config';
 import { handleControllerError } from '../../../lib/helpers';
 import { rateLimiterMiddleware } from '../../../lib/rate-limit';
-import { rejectRestrictedCountries } from '../../../lib/restrict-countries';
 import { CheckHistoricalBlockTradingRewardsSchema } from '../../../lib/validation/schemas';
 import { handleValidationErrors } from '../../../request-helpers/error-handler';
 import ExportResponseCodeStats from '../../../request-helpers/export-response-code-stats';
@@ -57,7 +56,6 @@ class HistoricalTradingRewardAggregationsController extends Controller {
 
 router.get(
   '/:address',
-  rejectRestrictedCountries,
   rateLimiterMiddleware(getReqRateLimiter),
   ...CheckHistoricalBlockTradingRewardsSchema,
   ...checkSchema({
