+++ indexer/services/comlink/src/controllers/api/v4/perpetual-positions-controller.ts
@@ -29,7 +29,7 @@ import {
 
 import { getReqRateLimiter } from '../../../caches/rate-limiters';
 import config from '../../../config';
-import { complianceCheck } from '../../../lib/compliance-check';
+import { complianceAndGeoCheck } from '../../../lib/compliance-and-geo-check';
 import { NotFoundError } from '../../../lib/errors';
 import {
   getFundingIndexMaps,
@@ -38,7 +38,6 @@ import {
   initializePerpetualPositionsWithFunding,
 } from '../../../lib/helpers';
 import { rateLimiterMiddleware } from '../../../lib/rate-limit';
-import { rejectRestrictedCountries } from '../../../lib/restrict-countries';
 import {
   CheckLimitAndCreatedBeforeOrAtSchema,
   CheckSubaccountSchema,
@@ -150,7 +149,6 @@ class PerpetualPositionsController extends Controller {
 
 router.get(
   '/',
-  rejectRestrictedCountries,
   rateLimiterMiddleware(getReqRateLimiter),
   ...CheckSubaccountSchema,
   ...CheckLimitAndCreatedBeforeOrAtSchema,
@@ -168,7 +166,7 @@ router.get(
     },
   }),
   handleValidationErrors,
-  complianceCheck,
+  complianceAndGeoCheck,
   ExportResponseCodeStats({ controllerName }),
   async (req: express.Request, res: express.Response) => {
     const start: number = Date.now();
