+++ indexer/services/comlink/src/lib/compliance-and-geo-check.ts
@@ -17,9 +17,17 @@ import { create4xxResponse } from './helpers';
 import { getIpAddr, isIndexerIp } from './utils';
 
 /**
- * Return an error code for users that access the API from a restricted country
+ * Checks if the address in the request is blocked or not.
+ *
+ * IF the address is in the compliance_status table and has the status CLOSE_ONLY,
+ * return data for the endpoint ELSE
+ * IF the address has compliance_status of BLOCKED block access to the endpoint (return 403) ELSE
+ * IF the origin country is restricted geography, block access to the endpoint (return 403) ELSE
+ * return data for the endpoint
+ * NOTE: This middleware must be used after `checkSchema` to ensure `matchData` can get the
+ * address parameter from the request.
  */
-export async function rejectRestrictedCountries(
+export async function complianceAndGeoCheck(
   req: express.Request,
   res: express.Response,
   next: express.NextFunction,
@@ -31,12 +39,7 @@ export async function rejectRestrictedCountries(
     return next();
   }
 
-  const {
-    address,
-  }: {
-    address: string,
-  } = matchedData(req) as AddressRequest;
-  console.log('address', address);
+  const { address }: AddressRequest = matchedData(req) as AddressRequest;
   const updatedStatus: ComplianceStatusFromDatabase[] = await ComplianceStatusTable.findAll(
     { address: [address] },
     [],
